requirements.txt
fastapi
uvicorn
python-jose
bcrypt
python-dotenv
pymongo
pydantic
python-multipart
python-jobspy

